+++ src/js/interfaces.js
@@ -1,3 +1,6 @@
 "use strict";
+/**
+ * @modified 2021-01-10 Added the `CanvasWrapper` interface.
+ **/
 Object.defineProperty(exports, "__esModule", { value: true });
 //# sourceMappingURL=interfaces.js.map